requirements.txt
torch
torchvision
transformers
streamlit
pandas
pillow
scikit-learn
numpy
